chatbot_service/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
google-generativeai==0.3.2
langchain>=0.1.0
aio-pika==9.3.1
redis>=5.0.1
httpx==0.25.2

# Vector Storage
chromadb>=0.4.18

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0 